Got FTS working on InMemoryProvider.  Refactored a bunch of FTS helper logic into a shared FullTextSearchHelpers module.
